fix(app): SSR 렌더링 오류 해결
빌드시에 SSR 관련하여 document 문제가 발생해서 ssr을 임시로 다 꺼둠. 하면서 경로 변경되어서 추가 수정
