+++ card-capture/src/app/page.tsx
@@ -1,20 +1,10 @@
-'use client';
+import dynamic from 'next/dynamic';
 
-import NavigationBar from '@/components/common/NavigationBar/NavigationBar';
-import React from 'react';
-import Banner from '@/components/main/Banner/Banner';
-import Description from '@/components/main/Description/Description';
-import TemplateGallery from '@/components/main/TemplateGallery/TemplateGallery';
-import SearchTemplate from '@/components/main/SearchTemplate/SearchTemplate';
+const MainPage = dynamic(() => import('./MainContent'), {
+  ssr: false,
+  loading: () => <p></p>,
+});
 
-export default function Home() {
-  return (
-    <div className="w-screen overflow-y-scroll font-Pretendard">
-      <NavigationBar isTransparent={true} />
-      <Banner />
-      <Description />
-      <TemplateGallery />
-      <SearchTemplate />
-    </div>
-  );
+export default function Page() {
+  return <MainPage />;
 }
