+++ card-capture/src/components/prompt/PromptInput/components/PromptModelInput/PromptModelInput.tsx
@@ -3,7 +3,7 @@ import PromptTitleText from '@/components/prompt/PromptInput/components/common/P
 import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import useIsMobile from '@/hooks/useIsMobile';
 import { UseFormSetValue } from 'react-hook-form';
-import { PromptInputFormType } from '@/app/prompt/page';
+import { PromptInputFormType } from '@/app/prompt/PromptContent';
 
 type PromptModelInputProps = {
   setValue: UseFormSetValue<PromptInputFormType>;
