+++ card-capture/src/components/prompt/PromptInput/components/PromptPurposeInput/PromptPurposeInput.tsx
@@ -1,7 +1,7 @@
 import PromptCategoryText from '@/components/prompt/PromptInput/components/common/PromptCategoryText';
 import PromptTitleText from '@/components/prompt/PromptInput/components/common/PromptTitleText';
 import { UseFormRegister } from 'react-hook-form';
-import { PromptInputFormType } from '@/app/prompt/page';
+import { PromptInputFormType } from '@/app/prompt/PromptContent';
 
 type PromptPurposeInputProps = {
   register: UseFormRegister<PromptInputFormType>;
